LayoutTests/webrtc/remote-track-label-quirks.html

<!doctype html>
<html>
    <head>
        <meta charset="utf-8">
        <script src="../resources/testharness.js"></script>
        <script src="../resources/testharnessreport.js"></script>
    </head>
    <body>
        <script src="routines.js"></script>
        <script>
promise_test(async (test) => {
    if (!window.internals)
        return;

    window.internals.setTopDocumentURLForQuirks("https://www.facebook.com");

    const localStream = await navigator.mediaDevices.getUserMedia({audio: true, video: true});
    test.add_cleanup(() => localStream.getTracks().forEach(track => track.stop()));
    const stream = await new Promise((resolve, reject) => {
        createConnections((firstConnection) => {
            firstConnection.addTrack(localStream.getAudioTracks()[0], localStream);
            firstConnection.addTrack(localStream.getVideoTracks()[0], localStream);
        }, (secondConnection) => {
            secondConnection.ontrack = (trackEvent) => {
                resolve(trackEvent.streams[0]);
            };
        });
        test.step_timeout(() => reject("Connections creation timed out"), 5000);
    });

    assert_equals(stream.getAudioTracks()[0].label, "remote audio - " + stream.getAudioTracks()[0].id, "remote audio track label");
    assert_equals(stream.getVideoTracks()[0].label, "remote video - " + stream.getVideoTracks()[0].id, "remote video track label");
});
        </script>
    </body>
</html>
